Create fractals_v1
This is the first version. It displays a Mandelbrot set and allows infinite zoom into any part of the images. By Infinite, I mean I have tried a few times and haven't reached the limit of the programs. I attribute this to python's ability to represent numbers of any precisions.